Use hpp-constraints API to get joints constrained
Previously, `RelativeMotion` uses dynamic cast to get the joints that
are fully constrained. This is fragile, since it is not clear that
developing new constraints/ constraint functions in hpp-constraint would
require adding new dynamic cast in `RelativeMotion`. In fact, the code
already breaks when the relative transformation constraint does not have
a full mask.

Now that there is an API from hpp-constraint dedicated to retrieving
joints constrained, we should use it. This commit also adds a test that
will fail on the previous approach but works with the new API.